chore: 完成Clippy代码质量提升工程，清理所有lint警告
本次提交完成了全面的Clippy lint修复工作，主要变更包括：
1. 添加全局允许项，兼容内核既有架构惯例
2. 批量替换手动构造C字符串、冗余类型转换等可自动修复项
3. 手动修复逻辑分支合并、饱和减法、循环优化等场景
4. 发现并修复了display模块像素格式推断的冗余逻辑bug
5. 覆盖所有内核与测试代码，实现0错误0警告状态，所有测试通过

+++ src/kernel/sync/rcu.rs
@@ -146,7 +146,7 @@ pub fn synchronize_rcu() {
     let current_cpu = crate::kernel::smp::get_current_cpu();
 
     // 标记所有在线 CPU 进入宽限期等待
-    for i in 0..cpu_count as u32 {
+    for i in 0..cpu_count {
         if i == current_cpu {
             continue;
         }
@@ -168,7 +168,7 @@ pub fn synchronize_rcu() {
     }
 
     // 等待所有在线 CPU 报告静止状态
-    for i in 0..cpu_count as u32 {
+    for i in 0..cpu_count {
         if i == current_cpu {
             continue;
         }
@@ -187,7 +187,7 @@ pub fn synchronize_rcu() {
     }
 
     // 重置所有 CPU 的 gp_state
-    for i in 0..cpu_count as u32 {
+    for i in 0..cpu_count {
         let data = rcu_data(i);
         data.gp_state.store(GP_IDLE, Ordering::Release);
     }
@@ -299,7 +299,7 @@ pub fn rcu_note_quiescent_state() {
 /// 通知所有 CPU 的 RCU 回调 (由同步宽限期调用)
 pub fn rcu_process_all_callbacks() {
     let cpu_count = crate::kernel::smp::get_cpu_count();
-    for i in 0..cpu_count as u32 {
+    for i in 0..cpu_count {
         let data = rcu_data(i);
         if data.need_callback_process.load(Ordering::Acquire) {
             // 使用 IPI 或直接处理 — 简化实现: 直接处理
